+++ src/client/modules/Reminders/Settings/InvestmentNoRecentInteractionForm.jsx
@@ -6,28 +6,28 @@ import { Form } from '../../../components'
 import { OPTION_YES } from '../../../../common/constants'
 import urls from '../../../../lib/urls'
 
+const redirectUrl = `${urls.reminders.settings.index()}?investments_no_recent_interaction=true`
+
 import {
-  TASK_GET_EXPORT_NRI_REMINDER_SUBSCRIPTIONS,
-  TASK_SAVE_EXPORT_NRI_REMINDER_SUBSCRIPTIONS,
+  TASK_GET_NRI_REMINDER_SUBSCRIPTIONS,
+  TASK_SAVE_NRI_REMINDER_SUBSCRIPTIONS,
 } from '../state'
 import {
   transformReminderDaysToAPI,
   transformNRIFormValuesToAnalyticsData,
 } from '../transformers'
 
-const redirectUrl = `${urls.reminders.settings.index()}?exports_no_recent_interactions=true`
-
-const ExportNoRecentInteractionForm = () => (
+const InvestmentNoRecentInteractionForm = () => (
   <Form
-    id="reminders-settings-export-no-recent-interaction"
-    initialValuesTaskName={TASK_GET_EXPORT_NRI_REMINDER_SUBSCRIPTIONS}
-    submissionTaskName={TASK_SAVE_EXPORT_NRI_REMINDER_SUBSCRIPTIONS}
+    id="reminders-settings-no-recent-interaction"
+    initialValuesTaskName={TASK_GET_NRI_REMINDER_SUBSCRIPTIONS}
+    submissionTaskName={TASK_SAVE_NRI_REMINDER_SUBSCRIPTIONS}
     redirectTo={() => redirectUrl}
     transformPayload={(formValues) => ({
       reminder_days: transformReminderDaysToAPI(formValues),
       email_reminders_enabled: formValues.emailNotifications === OPTION_YES,
     })}
-    analyticsFormName="editExportNoRecentInteractionReminderSettings"
+    analyticsFormName="editNoRecentInteractionReminderSettings"
     analyticsData={transformNRIFormValuesToAnalyticsData}
     flashMessage={() => 'Settings updated'}
     cancelRedirectTo={() => redirectUrl}
@@ -36,7 +36,7 @@ const ExportNoRecentInteractionForm = () => (
       <>
         {reminder_days && (
           <CommonNoInteractionFields
-            entity="exports"
+            entity="projects"
             reminder_days={reminder_days}
             email_reminders_enabled={email_reminders_enabled}
             reminders={reminders}
@@ -47,4 +47,4 @@ const ExportNoRecentInteractionForm = () => (
   </Form>
 )
 
-export default ExportNoRecentInteractionForm
+export default InvestmentNoRecentInteractionForm
